+++ .env.example
@@ -1,0 +1,1 @@
+GOOGLE_ANALYTICS_ID=G-XXXXXXXXXX
