+++ .gitignore
@@ -7,3 +7,5 @@ node_modules
 .data
 .env
 dist
+
+.vscode
