+++ components/common/VercelAnalytics.vue
@@ -1,5 +1,5 @@
 <template>
-    <!-- このコンポーネントは視覚的な出力を持ちません -->
+    <!-- 架空 -->
   </template>
   
   <script setup>
@@ -9,10 +9,10 @@
   onMounted(() => {
     // Vercelデプロイ環境のみで実行
     if (typeof window !== 'undefined' && window.location.hostname !== 'localhost' && !window.location.hostname.includes('192.168')) {
-      // Vercel Web Analyticsを初期化
+      // VercelWebAnalyticsを初期化
       inject();
       
-      // Vercel Speed Insightsを初期化
+      // VercelSpeedInsightsを初期化
       injectSpeedInsights();
       
       // 本番環境でのみスクリプトを動的に追加
@@ -24,7 +24,6 @@
         document.body.appendChild(script);
       };
       
-      // Vercelのスクリプトを追加
       addScript('/_vercel/insights/script.js');
       addScript('/_vercel/speed-insights/script.js');
     }
