+++ utils/constants.ts
@@ -1,0 +1,66 @@
+export const MENU_ITEMS = [
+    { name: 'ホーム', path: '/' },
+    { name: 'コミュニティについて', path: '/about' },
+    { name: '運営メンバー', path: '/team' },
+    { name: 'Pizzaトークン', path: '/token' },
+    { name: 'お問い合わせ', path: '/contact' }
+  ]
+  
+  export const SOCIAL_LINKS = {
+    discord: {
+      name: 'Discord',
+      url: 'https://discord.com/invite/BbxuwH5WS3',
+      icon: 'fab fa-discord'
+    },
+    twitter: {
+      name: 'X (Twitter)',
+      url: 'https://twitter.com/Web3studenttalk',
+      icon: 'fab fa-twitter'
+    },
+    note: {
+      name: 'note',
+      url: 'https://note.com/web3studentstalk',
+      icon: 'font-bold'
+    }
+  }
+  
+  export const TOKEN_LINKS = {
+    distribution: {
+      name: '公式ピザ配布基準',
+      url: 'https://west-pizza.notion.site/2d54b18c324e41c0af7af6bd8b4262b6'
+    },
+    guide: {
+      name: 'ピザのトリセツ',
+      url: 'https://west-pizza.notion.site/b0ed8ded66fd47fcb46acb6c03178c33'
+    }
+  }
+  
+  export const TYPEWRITER_PHRASES = [
+    '未来を創ろう', 
+    '課題を解決しよう', 
+    '可能性を広げよう', 
+    '価値を共有しよう'
+  ]
+  
+  export const FAQ_ITEMS = [
+    {
+      question: '参加に費用はかかりますか?',
+      answer: 'いいえ、コミュニティへの参加は完全無料です。どなたでも気軽にご参加いただけます。'
+    },
+    {
+      question: '学生以外でも参加できますか?',
+      answer: 'はい、社会人の方も参加可能です。Web3に興味がある方なら誰でも歓迎しています。'
+    },
+    {
+      question: '初心者でも大丈夫ですか?',
+      answer: 'もちろんです。むしろ初心者の方の参加を歓迎しています。気軽に質問できる環境を整えています。'
+    },
+    {
+      question: 'Xスペースはいつ開催されますか?',
+      answer: '基本的に毎週火・木・金曜日の夜に開催しています。詳しい日程はX (旧Twitter) やDiscordでお知らせしています。'
+    },
+    {
+      question: 'Pizzaトークンはどうすれば獲得できますか?',
+      answer: 'Xスペースへの参加やDiscordでの活動、コミュニティへの貢献などで獲得できます。詳しくはPizzaトークンのページをご覧ください。'
+    }
+  ]
